Fix __eq__ methods raising RecursionError
Override the `Ordering.__eq__` method, as the inherited supermethod
(from Mapping) compares both the keys and values. That is not only the
elements of the Ordering, but also their associated OrderingItems are
compared, which in turn requires the Orderings associated with said
OrderingItems to be compared, resulting in infinite recursion.

Tweak order of comparison in `OrderingItem.__eq__` to perform the
(slower) comparison of the `ordering` attributes if and only if the
(faster) comparison of the `item` attributes is True.

Remove unnecessary `bool()` call in `OrderingItem.__eq__`.

Add unit tests to for checking equality of Orderings or OrderingItems.