Improve errors raised by assert_* methods
Change `Ordering.assert_new_item` to raise a ValueError instead of a
KeyError, upon being called with a duplicate item.

Replace literal class name with dynamic lookup, so it will be correct
for subclasses

Avoid printing the potentially large repr of an Ordering to the
traceback, the traceback itself is enough to pinpoint which Ordering
is raising the exception.

Add the repr conversion flag to the formatting of the errant item, to
differentiate between e.g. 2 and '2'